requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn
openai
python-dotenv
scikit-learn